backend/requirements.minimal.txt
# MockMarket Backend - Core Dependencies
# Production-ready minimal requirements

# Web framework
Flask==3.1.2
flask-cors==5.0.0
Werkzeug==3.1.3

# Auth
flask-jwt-extended==4.7.1
google-auth==2.30.0

# Database
mysql-connector-python==9.2.0

# HTTP client
requests==2.32.5

# News feed parsing
feedparser==6.0.11

# Environment
python-dotenv==1.1.1

# Logging/utilities
blinker==1.9.0
click==8.3.0
colorama==0.4.6
rich==13.7.0

# Upstox SDK (optional, if using their client library)
# upstox-python-sdk==2.18.0
